requirements.txt
# Core
pandas
numpy
requests
python-dotenv
websocket-client
py-clob-client
pyyaml

# Data Storage
pyarrow